apps/hermes/requirements.txt
# Hermes dependencies
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────────────
httpx>=0.27,<1             # Async/sync HTTP — Ollama, Gemini, web fetch
pyyaml>=6.0,<7             # Config and context YAML parsing
rich>=13.0,<14             # CLI output formatting

# ── Phase 2 (uncomment when implementing) ─────────────────────────────────────
# google-generativeai>=0.8,<1    # Gemini SDK

# ── Phase 3 (uncomment when implementing) ─────────────────────────────────────
# python-telegram-bot>=21.0,<22  # Telegram async bot

# ── Phase 4 (uncomment when implementing) ─────────────────────────────────────
# fastapi>=0.115,<1              # Web UI backend
# uvicorn>=0.30,<1               # ASGI server
# jinja2>=3.1,<4                 # Web UI templates

# ── Second Brain integration (uncomment when Second Brain DB is available) ────
# psycopg2-binary>=2.9,<3        # Postgres driver (pgvector)

# ── Phase 5 (uncomment when n8n migration complete) ───────────────────────────
# mcp>=1.0,<2                    # MCP client library (n8n integration)

# ── Development / testing ─────────────────────────────────────────────────────
pytest>=8.0,<9
pytest-cov>=5.0,<6
